requirements.txt
python-dotenv>=1.0.0
streamlit>=1.32.2
PyMuPDF>=1.24.0
chromadb>=0.4.24
pyngrok>=7.1.5
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0